feat: add panic recovery testing and memory pressure protection
Addresses Claude Code review Priority 1.3 and Priority 2.2 concerns.

Changes:
1. Panic Recovery Test Coverage (Priority 2.2):
   - Added testPanicInjector hook for test-only panic injection
   - Implemented comprehensive TestWorkerPanicRecovery test
   - Verifies worker restarts after panic and processes subsequent requests
   - Validates done channel signaling during panic recovery

2. Memory Pressure Protection (Priority 1.3):
   - Added maxTrackersPerRevision limit (5000 pods)
   - Each tracker uses ~80KB, limit caps memory at ~400MB per revision
   - New metric: revisionTrackerLimitExceeded tracks rejections
   - Pod additions beyond limit are rejected with error logging
   - Comprehensive TestMemoryPressureProtection with configurable limit

3. Error Handling Documentation:
   - Documented why current error handling behavior is acceptable
   - State updates eventually consistent via K8s informer reconciliation
   - Dropped updates tracked via metrics for monitoring

All tests pass. This completes all Priority 1 and Priority 2 concerns
from the Claude Code review.

+++ pkg/queue/sharedmain/main.go
@@ -447,9 +447,9 @@ func buildBreaker(logger *zap.SugaredLogger, env config) *queue.Breaker {
 	// allow the autoscaler time to react.
 	queueDepth := 10 * env.ContainerConcurrency
 	params := queue.BreakerParams{
-		QueueDepth:      queueDepth,
-		MaxConcurrency:  env.ContainerConcurrency,
-		InitialCapacity: env.ContainerConcurrency,
+		QueueDepth:      uint64(queueDepth),
+		MaxConcurrency:  uint64(env.ContainerConcurrency),
+		InitialCapacity: uint64(env.ContainerConcurrency),
 	}
 	logger.Infof("Queue container is starting with BreakerParams = %#v", params)
 	return queue.NewBreaker(params)
